Update readme file to reflect a fix to a common issue I was facing
With zsh, unquoted `**/*.rb` was getting expanded and I was getting the error I explained above.
I have faced this for a long time, and with multiple languages, and only now did I buckle down and pry in to check the issue.
I wanted to share this with others who might be facing the same issue.