Fix for resume issue
The issue seems to be caused by a change of behavior within the watcher
and ultimately listen module.  Before, it was necessary to check the
internal state of the listen module from the runner module to be able to
resume from pause properly.  However, with the later versions of listen,
this doesn't seem necessary anymore and the decision whether to actually
resume or not can be done by the watcher module.